basic array/slice support
Break slices and arrays into two groups:
 - scalar-ish which get one column
 - struct-ish which have n columns (one per field)

Interface internal types are checked to see if they are uniformly
populated with the same types (or nil). If they are the same, we can
treat them the same as concrete types. Otherwise, we'll delegate and
probably end up with tables inside table-elements.